requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
blis==1.0.2
boto3==1.36.6
botocore==1.36.6
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
cymem==2.0.10
Cython==3.0.11
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
duckduckgo_search==7.2.1
durationpy==0.9
emoji==2.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eval_type_backport==0.2.2
executing==2.1.0
fastapi==0.115.5
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.3
gensim_sum_ext==0.1.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
json_repair==0.35.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kubernetes==31.0.0
langchain==0.3.14
langchain-aws==0.2.11
langchain-chroma==0.2.0
langchain-community==0.3.14
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-mistralai==0.2.4
langchain-neo4j==0.2.0
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.51
langsmith==0.2.10
language_data==1.3.0
listparser==0.20
llama-cloud==0.1.5
llama-index==0.12.2
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-chroma==0.4.0
llama-parse==0.5.15
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minify_html==0.15.0
mmh3==5.0.1
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.15.1
ndjson==0.3.1
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.19.2
openai==1.59.6
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
playwright==1.49.1
posthog==3.7.4
preshed==3.0.9
primp==0.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
Protego==0.3.1
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycorenlp==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==12.0.0
Pygments==2.18.0
PyMySQL==1.1.1
pyOpenSSL==24.3.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
RapidFuzz==3.11.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rq==2.0.0
rsa==4.9
s3transfer==0.11.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
scrapegraphai==1.37.0
Scrapy==2.12.0
semchunk==3.0.1
sentence-transformers==3.3.1
service-identity==24.2.0
setuptools==75.6.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
simpleeval==1.0.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.2
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
Twisted==24.10.0
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
undetected-playwright==0.3.0
unstructured==0.16.12
unstructured-client==0.28.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
w3lib==2.2.1
wasabi==1.1.3
watchfiles==1.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.0
yfiles_jupyter_graphs==1.10.1
youtube-transcript-api==0.6.3
zipp==3.21.0
zope.interface==7.2
